Refactored code

+++ Hands_On_List_1/19/19.c
@@ -13,7 +13,7 @@ Date: 31st Aug, 2024.
 #include <x86intrin.h>
 
 int main() {
-    uint64_t start, end;
+    unsigned long long start, end;
     pid_t pid;
 
     start = __rdtsc();
@@ -22,10 +22,10 @@ int main() {
 
     end = __rdtsc();
 
-    uint64_t cycles = end - start;
+    unsigned long long cycles = end - start;
 
     printf("Process ID: %d\n", pid);
-    printf("Time taken for getpid(): %lu cycles\n", cycles);
+    printf("Time taken for getpid(): %llu cycles\n", cycles);
 
     return 0;
 }
